+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - State merging via speculative execution of both branches of a JUMPI, joining the
   resulting states whenever possible. Amount of speculative execution is
   controlled via `merge-max-budget`
+- Missing simplifications for Eq, Mod, SMod, XOR, SHL, SHR, and Or
 
 ## Changed
 - Simplifier now rewrites `Mul(-1, x)` and `~x + 1` to `Sub(0, x)`
@@ -26,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The option `--smttimeout` is now called `--smt-timeout` for consistency with other
   options.
 - We now abort VM run on a failed `assume`.
+- Removed SHA256 from the Expr since it was not being used
 
 ## [0.57.0] - 2026-01-08